+++ routes/web.php
@@ -48,7 +48,7 @@ Route::get('/dashboard', Feed::class)
 Route::middleware('auth')->group(function () {
     Route::get('/profile', Profile::class)->name('profile');
     Route::get('/publish', PublishWorkflow::class)->name('publish');
-    Route::get('/groups', GroupManager::class)->name('groups');
+    Route::get('/groups', GroupManager::class)->middleware('karma:group.create')->name('groups');
     Route::get('/posts/{postId}', PostDetail::class)->name('posts.detail');
     Route::get('/posts/{postId}/update', UpdatePost::class)->name('posts.update');
     Route::get('/settings', [ProfileController::class, 'edit'])->name('profile.edit');